argo_workflow/seldon/model/requirements.txt
seldon-core
determined==0.12.13
torch
Pillow
numpy
scikit-image
torchvision
python_pachyderm
